+++ skoolkit/config.py
@@ -24,6 +24,7 @@ COMMANDS = {
     'sna2skool': {
         'Base': (10, 'base'),
         'Case': (2, 'case'),
+        'CommentWidthMin': (10, ''),
         'DefbMod': (1, ''),
         'DefbSize': (8, ''),
         'DefbZfill': (0, ''),